Move Loop renaming into own stage
